+++ lib/vagrant/errors.rb
@@ -364,6 +364,10 @@ module Vagrant
       error_key(:not_found, "vagrant.actions.vm.host_only_network")
     end
 
+    class NFSBadExports < VagrantError
+      error_key(:nfs_bad_exports)
+    end
+
     class NFSCantReadExports < VagrantError
       error_key(:nfs_cant_read_exports)
     end